Problem 3 solved

+++ euler/src/main/java/problem3/PrimeFactors.java
@@ -2,10 +2,10 @@ package problem3;
 
 public class PrimeFactors {
 
-    public static Integer generateMax(int n) {
-        int maxPrimeFactor = 0;
+    public static Long generateMax(long n) {
+        long maxPrimeFactor = 0;
 
-        for (int candidate = 2; n > 1; candidate++) {
+        for (long candidate = 2; n > 1; candidate++) {
             for (; n % candidate == 0; n /= candidate) {
                 maxPrimeFactor = candidate;
             }
